+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### 0.6a
 
 - Flow Path component outputs paths as a flat list; not a tree
+- Fixed generic plant attrbiutes
+- Restructured directory
 - Switched references to use NuGet
 
 #### 0.5a