+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 # Changelog
 
+<!--
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+-->
 
-## [Unreleased]
+## [0.6.1a] - 2017-11-16
+### Changed
+- Improved component Icons slightly (proper icons coming soon!)
 
-## [0.6.0a]
+### Removed
+- Plant Placement Solver component (needs to be reworked form the C# script)
+
+## [0.6.0a] - 2017-08-27
 ### Changed
 - Flow Path component outputs paths as a flat list; not a tree
 - Flow Path component now split into seperate versions for Surfaces and Meshes
@@ -21,3 +28,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Initial release for Skin + Scale Studio
 - Tidied up example definitions
+